move to internal/automation package
